pyproject.toml
[project]
name = "traininfo-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bottle>=0.13.4",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "schedule>=1.2.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
